Convert global_settings to new custom resource format
